Finish extracting maps into separate component.
